This commit changes how the log likelihood is calculated.
Instead of using the simplified recursive equations present in the original
paper, this commit changes the calculation of the loglikelihood to use the
beta functions directly.

This change dramatically impact training time since it can leverage the full
potential of optimized array calculations offered by scipy and numpy.

Additionally a small bug with python 3's format is fixed as well as minor style
changes.