House cleaning and analysis tools
This commit fixes files names, import paths and a number of other small
changes. Additionally it also removes some legacy python2 campatibility
code that is no longer necessary. Finally is makes a number of small
changes here and there as well as make considerable inroads in an
analysis using this particular company data.

One important fix was the issue of zero age. Instead of ever returning
zero ages (people for whom upgrade and cacelation date happened in the
same month), we instead return their age as being one. This is in line
with how the original paper formulates the problem and not only avoids a
number of weird issues but also leads to substantially better results.

Misc commit messages:
- move example data to main package folder
- remove notebook added by mistake
- partial refactor work
- trial run of model
- partial exploratory work
- change output of predict function to share index and stack data
- progress with analysis, retention curves, ltv calculations, everything
- stratifying analysis, and a bit of clean up
- facts
- Formalize generation of standard working data frames with data handler
- continue with facts
- Fix a number of little and big buggs
- subscription data ready for testing