Organize analysis pipeline
This commit does a lot of different things, most of which are intended
to organize the analysis pipeline.

Important changes are:
- Files were organized. DataHandler was moved to the util file, together
with SubscriptionData. The Loader classes were put in their own,
loader.py, file.
- Loading data scrip. A simple script to load data was added. It pickles
the data object when is done, avoiding the need to parse the data over
and over, which can be time consuming.
- .pkl files were added to the gitignore
- Analysis notebooks were broken up by actuals, traditional, and sbg.
They were also cleaned in the process.

This is still a work in progress.

PART II
This commit also formalizes how train and test split are carried out by
the SubscriptionData object, specially when it comes to spliting by
date. This is now done in such a way that date splits leads to no
leakage of data, allowing for more trustworthy validation.